mcp_agent/orchestrator/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure SDKs
azure-identity==1.15.0
azure-cosmos==4.5.1
openai==1.3.0

# HTTP client
httpx==0.25.2

# JWT handling
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
